Move load to elements.py
